Encapsulate option parsing
Moves option parsing methods out of main class to improve clarity
Cleanups up some private method testing and hackery in main class tests
